moving feature level components out of the compoentns into features folder for clearer segregation of concerns
