+++ frontend/src/features/uploadsection/UploadPhoto.tsx
@@ -1,6 +1,6 @@
 
 import React, { useEffect } from 'react';
-import { useFeedbackContext } from '../contexts/FeedbackContext';
+import { useFeedbackContext } from '../../components/contexts/FeedbackContext'; 
 
 const UploadPhoto: React.FC = () => {
     const { setImageUrl, setFeedback, setLoading } = useFeedbackContext();
