+++ frontend/src/features/uploadsection/UploadSection.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import UploadPhoto from './UploadPhoto';
 import ImageCard from '../feedback/ImageCard';
-import { useFeedbackContext } from '../contexts/FeedbackContext';
+import { useFeedbackContext } from '../../components/contexts/FeedbackContext';
 
 const UploadSection: React.FC = () => {
     const {imageUrl} = useFeedbackContext();
